conclud problem 3 of CSES

+++ CSES/2-missing-number.c
@@ -1,11 +1,12 @@
 #include <stdio.h>
 
-int main() {
+int main() 
+{
     int n;
     scanf("%d", &n);
 
-    long long expected_sum = (long long)n * (n + 1) / 2;
-    long long actual_sum = 0;
+    int expected_sum = n * (n + 1) / 2;
+    int actual_sum = 0;
     int x;
 
     for (int i = 0; i < n - 1; i++) {
@@ -13,7 +14,7 @@ int main() {
         actual_sum += x;
     }
 
-    printf("%lld\n", expected_sum - actual_sum);
+    printf("%d\n", expected_sum - actual_sum);
 
     return 0;
 }
